Updated the master base template saas-bootcamp-baseline-master.template to invoke 2 new series of templates to accelerate Lab 1 by provisioning the pieces from Parts 2-6.
Updated most of the template descriptions so the CloudFormation console
UI is a little easier to follow. Still have more descriptions to update.

The tenant, auth and registration deployments now follow the pattern of
the user service - copying the Node application over to an S3 bucket
which is the source for a CodePipeline build that does the Docker build
and Docker push to our existing ECR. A CFN template is then lit off by
Lambda which creates the ECS Service. The ESC Task definitions already
exist from the baseline deployments.