Removing the old PDF docs for Labs 1 and 2.
Updating the S3 bucket for the web client to enable web hosting
automatically.

Updating the Cloud9 environment to be owned by the assumed "player" role
for compatibility with the Event Engine. Note that this will break
deployment in a local account. We need to integrate some sort of
CloudFormation condition to alternate running in the EE or not.